moved payment classes into their own module
