+++ juntagrico_billing/admin/payment_inline.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from juntagrico_billing.entity.bill import Payment
+from juntagrico_billing.entity.payment import Payment
 
 
 class PaymentInline(admin.TabularInline):
