+++ juntagrico_billing/models.py
@@ -1,6 +1,7 @@
 from django.db.models import signals
 
-from juntagrico_billing.entity.bill import Payment, BillItem
+from juntagrico_billing.entity.bill import BillItem
+from juntagrico_billing.entity.payment import Payment
 from juntagrico_billing.lifecycle.payment import payment_saved
 from juntagrico_billing.lifecycle.billitem import billitem_saved
